+++ forui/test/src/widgets/time_field/picker/picker_time_field_test.dart
@@ -3,7 +3,6 @@ import 'package:flutter/widgets.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:forui/src/widgets/picker/picker_wheel.dart';
-import 'package:intl/date_symbol_data_local.dart';
 import 'package:intl/intl.dart';
 
 import 'package:forui/forui.dart';